Move functions and components to their own files
